+++ src/main/java/at/porscheinformatik/seleniumcomponents/SeleniumComponent.java
@@ -2,7 +2,6 @@ package at.porscheinformatik.seleniumcomponents;
 
 import java.util.Objects;
 
-import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebElement;
 
 /**
@@ -45,28 +44,6 @@ public interface SeleniumComponent extends WebElementContainer
      * @return true if the component exists
      */
     boolean isReady();
-
-    /**
-     * Returns true if a {@link WebElement} described by this component is visible. By default it checks, if the element
-     * is visible. This method has no timeout, it does not wait for the component to become existent.
-     *
-     * @return true if the component is visible
-     */
-    default boolean isVisible()
-    {
-        try
-        {
-            return SeleniumUtils.retryOnStale(() -> {
-                WebElement element = element();
-
-                return element.isDisplayed();
-            });
-        }
-        catch (NoSuchElementException e)
-        {
-            return false;
-        }
-    }
 
     @Override
     String toString();